move utils to separate folder
