+++ go/controllers/rayjob/controller_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/michelangelo-ai/michelangelo/go/controllers/utils/testutils"
 	apipb "github.com/michelangelo-ai/michelangelo/proto/api"
 	v2pb "github.com/michelangelo-ai/michelangelo/proto/api/v2"
 	v1 "github.com/ray-project/kuberay/ray-operator/apis/ray/v1"
@@ -245,7 +246,7 @@ func TestReconciler_Reconcile(t *testing.T) {
 			fakeRayV1Client := &rayv1fake.FakeRayV1{
 				Fake: &k8stesting.Fake{},
 			}
-			fakeClientWrapper := NewFakeClientWrapper(fakeClient)
+			fakeClientWrapper := testutils.NewFakeClientWrapper(fakeClient)
 
 			r := &Reconciler{
 				Client:         fakeClientWrapper,
