+++ go/controllers/utils/testutils/fake_client_wrapper.go
@@ -1,4 +1,4 @@
-package rayjob
+package testutils
 
 import (
 	"context"
